+++ wemake_python_styleguide/visitors/ast/functions.py
@@ -6,6 +6,7 @@ from typing import ClassVar, Dict, List, Optional, Union
 
 from typing_extensions import final
 
+from wemake_python_styleguide.compat.aliases import FunctionNodes
 from wemake_python_styleguide.constants import (
     FUNCTIONS_BLACKLIST,
     UNUSED_VARIABLE,
@@ -26,14 +27,13 @@ from wemake_python_styleguide.violations.refactoring import (
     UselessLambdaViolation,
     WrongIsinstanceWithTupleViolation,
 )
-from wemake_python_styleguide.visitors.base import BaseNodeVisitor
-from wemake_python_styleguide.visitors.decorators import alias
+from wemake_python_styleguide.visitors import base, decorators
 
 LocalVariable = Union[ast.Name, ast.ExceptHandler]
 
 
 @final
-class WrongFunctionCallVisitor(BaseNodeVisitor):
+class WrongFunctionCallVisitor(base.BaseNodeVisitor):
     """
     Responsible for restricting some dangerous function calls.
 
@@ -62,7 +62,7 @@ class WrongFunctionCallVisitor(BaseNodeVisitor):
 
     def _ensure_super_context(self, node: ast.Call) -> None:
         parent_context = nodes.get_context(node)
-        if isinstance(parent_context, (ast.FunctionDef, ast.AsyncFunctionDef)):
+        if isinstance(parent_context, FunctionNodes):
             grand_context = nodes.get_context(parent_context)
             if isinstance(grand_context, ast.ClassDef):
                 return
@@ -110,11 +110,11 @@ class WrongFunctionCallVisitor(BaseNodeVisitor):
 
 
 @final
-@alias('visit_any_function', (
+@decorators.alias('visit_any_function', (
     'visit_AsyncFunctionDef',
     'visit_FunctionDef',
 ))
-class FunctionDefinitionVisitor(BaseNodeVisitor):
+class FunctionDefinitionVisitor(base.BaseNodeVisitor):
     """Responsible for checking function internals."""
 
     _allowed_default_value_types: ClassVar[AnyNodes] = (
@@ -203,7 +203,7 @@ class FunctionDefinitionVisitor(BaseNodeVisitor):
 
 
 @final
-class UselessLambdaDefinitionVisitor(BaseNodeVisitor):
+class UselessLambdaDefinitionVisitor(base.BaseNodeVisitor):
     """This visitor is used specifically for ``lambda`` functions."""
 
     def _have_same_kwarg(self, node: ast.Lambda, call: ast.Call) -> bool: